requirements.txt
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
cattrs==1.10.0
certifi==2021.10.8
cfscrape==2.1.1
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
pytz==2021.3
requests==2.27.1
requests-cache==0.9.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
url-normalize==1.4.3
urllib3==1.26.8
